Performance improvement - call resize only when necessary.
A user of this library specifies the dimensions of the chart in their
render call. That means that every time their render function is
called, `eval SetDimensions ...` is invoked, which in turn calls echarts'
`resize` function, which is a _relatively_ expensive call.

This change ensures that resize is only triggered if the dimensions of
the chart have actually changed.